push-setup script for aiding manual testing
This is the start of a simple script + cookbook for setting up
a push-server and push-client cluster to help make manual push-jobs
testing a little quicker. It has similar goals to [analytics-setup](https://github.com/chef/chef-analytics/tree/master/dev/cookbooks/analytics-setup).

It uses `test-kitchen` and `knife` to stand up a chef-server VM
that has push-jobs-server installed and configured on it and then
also a chef-client VM that has the push-client installed and
configured on it.